+++ R/Ternary-package.R
@@ -1,0 +1,12 @@
+#' @details For usage instructions, please consult the guidance article 
+#' [online](https://ms609.github.io/Ternary/articles/Ternary.html) or by
+#' typing `vignette('Ternary', 'Ternary')` at the R console.
+#' 
+#' @keywords internal
+"_PACKAGE"
+
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
+## usethis namespace: start
+## usethis namespace: end
+NULL